Force the use of the JDK 11 HTTP Client
Okhttp used to be the only alternative in native but now the JDK 11 HTTP
client is supported and works perfectly in native.
By refactoring this support, they introduced the usage of a Okhttp
method that isn't present in the version enforced by the Quarkus BOM so
for now, we disable Okhttp entirely.

Also added some basic integration tests to test native support. I used
Quarkus GitHub App to test this until now but let's clean things up.

Fixes #113